ai/requirements.txt
tensorflow  
numpy
pandas
jupyter
neptune-client
redis
